% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abhgeno-data.R
\docType{data}
\name{abhgeno}
\alias{abhgeno}
\title{Rice F2 genotype data}
\format{
An object of class \code{"list"}; imported with \code{readABHgenotypes} from \code{ABHgenotypeR}.
}
\usage{
data(abhgeno)
}
\description{
Genotypic data from a QTL mapping experiment
Rice, with data on 200 F2 lines (IR64 x ASD1) and 3436 SNP markers.
}
\examples{
data(abhgeno)
j <- imputeByFlanks(abhgeno)
h <- correctUndercalledHets(j, 10)
k <- correctStretches(h, 7)
l <- maskTransitionSites(k, 4)
m <- maskChromosomeEnds(l, 4)
}
\keyword{datasets}
