% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskChromosomeEnds.R
\name{maskChromosomeEnds}
\alias{maskChromosomeEnds}
\title{Masking of double recombinations at chromosome ends}
\usage{
maskChromosomeEnds(inputGenos, maxHapLength = 1)
}
\arguments{
\item{inputGenos}{ABH genotype file to mask.}

\item{maxHapLength}{Max length (in number of markers) of double recombination haplotype to set to missing. Default is 1.}
}
\description{
Sets double recombination calls at the start and end of chromosomes to missing, e.g. in the chromosome start sequence of 5'-AAHH or end sequence BBAA-3', both AAs are replaced by NN.
}
\examples{
data(abhgeno)
g <- maskChromosomeEnds(abhgeno, 4)
g

}
\keyword{mask}
