% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskTransitionSites.R
\name{maskTransitionSites}
\alias{maskTransitionSites}
\title{Masking of double recombinations at homozygous-heterozygous transition sites}
\usage{
maskTransitionSites(inputGenos, maxHapLength = 1)
}
\arguments{
\item{inputGenos}{ABH genotype file to mask.}

\item{maxHapLength}{Max length (in number of markers) of double recombination haplotype to set to missing. Default is 1.}
}
\description{
Sets short double recombination calls to missing, e.g. in the sequence HHAABB, AA is replaced by missing data, NN.
}
\examples{

# replaces double recombination transitions up to 4 marker lengths
data(abhgeno)
g <- maskTransitionSites(abhgeno, 4)
g

}
